pyproject.toml
[tool.poetry]
name = "giphy-app"
version = "0.1.0"
description = "A quick app to display a gif from giphy"
authors = ["Nate Brengle nate.k.mooney@gmail.com"]

[tool.poetry.dependencies]
python = "^3.7"
argcomplete = "^1.10"
requests = "^2.22"
cytoolz = "^0.10.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
